+++ src/ffi/cpython/mod.rs
@@ -1,7 +1,9 @@
 pub mod abstract_;
 pub mod bytesobject;
 pub mod ceval;
+pub mod code;
 
 pub use self::abstract_::*;
 pub use self::bytesobject::*;
 pub use self::ceval::*;
+pub use self::code::*;
